octo-hub-node/pyproject.toml
[project]
name = "octo-hub-node"
version = "0.1.0"
description = "OctoHub Node Client - WebSocket client for task processing"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "websockets>=12.0",
    "aiohttp>=3.9.0",
    "asyncio-mqtt>=0.16.0",
    "python-dotenv>=1.0.0",
]
